Update and findAll implementation

+++ src/model/dao/impl/DepartmentDaoJDBC.java
@@ -5,6 +5,7 @@ import model.dao.DepartmentDao;
 import model.entities.Department;
 
 import java.sql.*;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -50,27 +51,42 @@ public class DepartmentDaoJDBC implements DepartmentDao {
                 DB.closeStatement(st);
                 DB.closeResultSet(rs);
             }
-
         }
 
         @Override
         public void update(Department obj) {
-
+            PreparedStatement st = null;
+            try {
+                if (obj.getId() != null) {
+                st = conn.prepareStatement("UPDATE department SET Name = ? WHERE id = ?");
+                st.setString(1, obj.getName());
+                st.setInt(2, obj.getId());
+                st.executeUpdate();
+                System.out.println("Update completed");
+                }
+            } catch (SQLException e) {
+                throw new DbException(e.getMessage());
+            } finally {
+                DB.closeStatement(st);
+            }
         }
 
         @Override
         public void deleteById(Integer id) {
             PreparedStatement st = null;
             try {
-                st = conn.prepareStatement("DELETE FROM department WHERE Id = ?");
-                st.setInt(1, id);
-                st.executeUpdate();
+                Department dep = findById(id);
+                if (dep != null) {
+                    st = conn.prepareStatement("DELETE FROM department WHERE Id = ?");
+                    st.setInt(1, id);
+                    st.executeUpdate();
+                    System.out.println("Delete completed");
+                }
             } catch (SQLException e) {
                 throw new DbException(e.getMessage());
             } finally {
                 DB.closeStatement(st);
             }
-
         }
 
         @Override
@@ -125,6 +141,26 @@ public class DepartmentDaoJDBC implements DepartmentDao {
 
         @Override
         public List<Department> findAll() {
-            return List.of();
+            PreparedStatement st = null;
+            ResultSet rs = null;
+            try {
+                st = conn.prepareStatement("SELECT * FROM department ORDER BY Name");
+                rs = st.executeQuery();
+
+                List<Department> list = new ArrayList<>();
+                while (rs.next()) {
+                    Department dep = new Department();
+                    dep.setId(rs.getInt("Id"));
+                    dep.setName(rs.getString("Name"));
+                    list.add(dep);
+                }
+                return list;
+            } catch (SQLException e) {
+                throw new DbException(e.getMessage());
+            } finally {
+                DB.closeStatement(st);
+                DB.closeResultSet(rs);
+            }
         }
+
 }
